+++ __tests__/array.test.ts
@@ -1,4 +1,5 @@
 import { unique } from '../src/array/unique'
+import { uniqueValue } from '../src/array/uniqueValue'
 
 describe('unique', () => {
   test('numbers', () => {
@@ -23,5 +24,23 @@ describe('unique', () => {
   })
   test('empty', () => {
     expect(unique([])).toStrictEqual([])
+  })
+})
+
+describe('uniqueValue', () => {
+  test('no mutation', () => {
+    expect(uniqueValue('e', ['a', 'b', 'c', 'd'])).toStrictEqual('e')
+    expect(uniqueValue('a', ['b', 'c', 'd'])).toStrictEqual('a')
+    expect(uniqueValue('a', [])).toStrictEqual('a')
+  })
+  test('mutation', () => {
+    expect(uniqueValue('a', ['a', 'b', 'c', 'd'])).toStrictEqual('a0')
+    expect(uniqueValue('a', ['a'])).toStrictEqual('a0')
+    expect(uniqueValue('a', ['a', 'a0', 'b', 'c', 'd'])).toStrictEqual('a1')
+    expect(uniqueValue('a', ['a', 'a0', 'a1', 'a2', 'a03'])).toStrictEqual('a3')
+  })
+  test('empty', () => {
+    expect(uniqueValue('', [])).toStrictEqual('')
+    expect(uniqueValue('a', [])).toStrictEqual('a')
   })
 })