+++ src/index.ts
@@ -5,6 +5,7 @@ export { truncate } from './text/truncate'
 
 // array
 export { unique } from './array/unique'
+export { uniqueValue } from './array/uniqueValue'
 
 // files
 export { getExtension } from './files/getExtension'