+++ CHANGELOG.md
@@ -5,12 +5,24 @@ Contains all modifications to the project.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2022-11-07 - cbs
+
+### Added
+
+`uniqueValue(value:string,list:string[])` for strings
+
+## [0.0.4] - 2022-07-11 - cbs
+
+### Added
+
+`truncate()` for strings
+
 ## [0.0.3] - 2022-07-06 - cbs
 
 ### Added
 
-getExtension() for files
-unique() for arrays
+`getExtension()` for files
+`unique()` for arrays
 
 ## [0.0.2] - 2022-07-05 - cbs